TS-21 Translate IMUSANT internal representations back to Music XML including bug fix TS TS-29 Error in xmlvisitor print function

+++ catsmat/catsmat/CounterpointAnalysis/src/CATSMAT_cp_matrix.cpp
@@ -275,8 +275,8 @@ namespace CATSMAT
     {
         if (fCurrentTime!=time && time!=IMUSANT_time()) {
             fCurrentTime = time;
-            vector<long> nums = fCurrentTime.getNumerator();
-            vector<long> doms = fCurrentTime.getDenominator();
+            vector<int> nums = fCurrentTime.getNumerator();
+            vector<int> doms = fCurrentTime.getDenominator();
             vector<TRational> rats;
             if (doms.size()<2) {
                 for (auto i = doms.begin(); i!=doms.end();i++) {
